Adds ability to apply facets in GraphQL
Our initial GraphQL implementation was intended as experiemental so it did not incldue the ability to apply facets.

As we start to consider whether we will move to GraphQL fully or stay with REST, it seemed worthwhile to consider how working with facets would feel in GraphQL both in terms of consuming the data and writing the code.

A separate commit will follow with tests but I wanted to get some feedback as-is.

https://mitlibraries.atlassian.net/browse/DISCO-60